Add a pylt.fail(string) which raises a PyltError() exception
Give ->rd() and ->ask() an optional "fail" argument, to control
if we want an exception or merely information about failure.

If a usbtmc bulk in fails, clear the pipes, ->fail().

Always config u2004a, otherwise it whines.

Move prologix device debug output to the per-prologix debug files.

(Typo in README)